+++ frontend/app/pages/groups/[id]/activity.vue
@@ -1,0 +1,11 @@
+<template>
+  <main v-if="group">
+    <h1>activity</h1>
+  </main>
+</template>
+
+<script setup lang="ts">
+const { group } = useDetailPageGroup();
+</script>
+
+<style scoped lang="scss"></style>
